增加权限控制

+++ fc/service/src/main/java/cn/yunovo/iov/fc/service/impl/CcOrgServiceImpl.java
@@ -278,6 +278,20 @@ public class CcOrgServiceImpl extends ServiceImpl<ICcOrgMapper, CcOrg> implement
 		return p;
 		
 	}
+	
+	@Override
+	public boolean permission(Integer org_id, String orgpos) {
+		
+		if(org_id == null || org_id < 0 || StringUtils.isEmpty(orgpos)) {
+			return false;
+		}
+		
+		if(StringUtils.equals("*", orgpos)) {
+			return true;
+		}
+		
+		return StringUtils.contains(","+orgpos+",", ","+org_id+",");
+	}
 
 	@Override
 	public int insert(OrgForm form, LoginInfo info) {
@@ -287,6 +301,13 @@ public class CcOrgServiceImpl extends ServiceImpl<ICcOrgMapper, CcOrg> implement
 		if(user == null) {
 			throw new BusinessException("无效操作");
 		}
+		
+		String orgpos = this.getOrgpos(user.getOrg_id(), user.getOrgpos());
+		
+		if(!this.permission(form.getParent_id(), orgpos)) {
+			throw new BusinessException("抱歉您无权限在此父机构下创建所属子机构!");
+		}
+		
 		CcOrg ccOrg = new CcOrg();
 		BeanUtils.copyProperties(form, ccOrg);
 		ccOrg.setPartner_id(RandomStringUtils.randomAlphanumeric(15));
@@ -306,7 +327,15 @@ public class CcOrgServiceImpl extends ServiceImpl<ICcOrgMapper, CcOrg> implement
 			throw new BusinessException("无效操作");
 		}
 		
-		//TODO 判断是否有删除权限
+		String orgpos = this.getOrgpos(user.getOrg_id(), user.getOrgpos());
+		
+		if(!this.permission(org.getOrg_id(), orgpos)) {
+			throw new BusinessException("系统提示：您无权限修改该机构信息");
+		}
+		
+		if(!this.permission(org.getParent_id(), orgpos)) {
+			throw new BusinessException("抱歉您无权限在此父机构下创建所属子机构");
+		}
 		
 		CcOrg ccOrg = new CcOrg();
 		BeanUtils.copyProperties(org, ccOrg);
@@ -318,9 +347,30 @@ public class CcOrgServiceImpl extends ServiceImpl<ICcOrgMapper, CcOrg> implement
 	}
 
 	@Override
-	public int delete(Integer[] orgs, LoginInfo loginBaseInfo) {
+	public int delete(Integer[] orgs, LoginInfo info) {
 		
-		//TODO 判断是否有删除权限
+		CcUser user = iCcUserService.findUserOrgAndOrgpos(info.getLoginName());
+		
+		if(user == null) {
+			throw new BusinessException("无效操作");
+		}
+		
+		if(orgs == null || orgs.length < 1) {
+			throw new BusinessException("请选择您需要删除的机构");
+		}
+		
+		String orgpos = this.getOrgpos(user.getOrg_id(), user.getOrgpos());
+		for (Integer o : orgs) {
+			
+			if(!this.permission(o, orgpos)) {
+				throw new BusinessException("系统提示：您无权删除该机构");
+			}
+			
+			if(o == user.getOrg_id()) {
+				
+				throw new BusinessException("系统提示： 您不能删除自己所属机构！");
+			}
+		}
 		
 		return iCcOrgMapper.deleteBatchIds(Arrays.asList(orgs));
